coverage: user can avoid distributing '.info' pages
Can be done like this:

        AUTOMAKE_OPTIONS = info-in-builddir
        dist-info:
            @:

Note that this usage is not yet documented: we might decide to go
for a fully-fledged 'no-dist-info' flag, or something like that, in
future automake version (this is not yet decided); in which case,
it's better not to have people start to rely on the hack above.
Still, there's no good reason to break it gratuitously, hence this
test coverage.

* t/txinfo-nodist-info.sh: New test.
* t/list-of-tests.mk: Add it.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>