texi: overriding of 'dist-info' target is broken
Due to the differences in the codebase of Automake-NG and mainline
Automake, that (undocumented) capability has been lost.  Not a big
deal, since in the long term, we should give a proper API to obtain
that effect, instead of forcing the user to employ and undocumented
hack.  So, until we have decided on that API, just mark the test
checking for the lost capability as "XFAIL."

* Makefile.am (XFAIL_TESTS): Add 't/txinfo-nodist-info.sh'.
* NG-NEWS: Document the regression.  Related adjustments and
updated.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>