depend: a preparatory refactoring
This is just a preparatory change in view of a future refactoring.
No semantic change is intended.

* lib/am/depend2.am: Adjust; for example, using '$@' in all recipes
to indicate the output file (rather than recipe-specific transforms
like '%OBJ%' and '%LTOBJ%'), and not assuming to know the exact
file extension of the output file (e.g., whether it's '.o' or '.lo').

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>