DGJPP: re-introduce support, on Windows 2000 and later
See automake bug#13435.

The problematic DGJPP-related code should no longer be necessary today:

 * the code to support DJGPP-style split info files ('*.i[0-9][0-9]')
   is no longer needed, since we have dropped support for split info
   files anyway (see commit v1.13.1-60-gcdba479 and automake  bug#13351);

 * on Windows 2000 and later, modern version of DJGPP support file names
   starting with dots.

With that, the remaining pieces of code required to support DJGPP on
non-ancient Windows are few and unobtrusive enough that we re-introduce
them, in order to keep DJGPP alive -- the cost/benefit ratio has become
definitely small enough for that.

Note that support for DJGPP on DOS and Windows 95/98/ME is *not*
reintroduced.  That is no longer worth worrying about.

For reference, here are the commits where we removed DJGPP support:
- v1.13-6-gad08bbf, "Drop support for DJGPP, MS-DOS, Windows 95/98/ME"
- v1.13-7-gff0c7f8, "general: assume dot-starting file names are supported"

* NEWS: Make clear that DJGPP on Windows 2000 and later should still be
supported.
* automake.in (BEGIN): Override $ENV{SHELL} for DJGPP.
* maintainer/syntax-checks.mk (automake_diff_no): Adjust, as there are
now eight (not just seven) different lines between 'automake.in' and
'automake'.
* bootstrap.sh ($BOOTSTRAP_SHELL): Give a more proper default for DJGPP.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>