NEWS: IRIX is still supported; only SGI depmode no longer is
This change is for the master branch (see the similar commit
'v1.13.1b-3-g97f04a5' on the maint branch).

The only IRIX-specific support that is actually removed in master (and
thus in the next major Automake version) is the depcomp support for the
SGI compiler.  That means that automatic dependency tracking will no
longer work with that compiler, but "normal" compilation should still
work, at least until the compiler is supported by Autoconf.

So there is no point in alarming our users by stating in the NEWS file
that "support for IRIX and the SGI compilers has been removed"; after
all, while we don't test on nor particularly care about IRIX anymore,
that doesn't mean we are deliberately breaking it, and the likelihood
of an intended breakage there is very low.

See:
<http://lists.gnu.org/archive/html/automake-patches/2013-01/msg00164.html>
<http://lists.gnu.org/archive/html/automake-patches/2013-01/msg00165.html>

* NEWS: Adjust.  Improve explanation of our rationale for dropping
support for the SGI compilers.

Suggested-by: Peter Rosin <peda@lysator.liu.se>
Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>