tests on TAP: don't run the driver with perl unconditionally
* t/ax/tap-setup.sh: When a 'Makefile.am' was pre-existent in the
test directory at the moment the client test script sourced this
file, said 'Makefile.am' was tweaked to provide it with a proper
definition of TEST_LOG_DRIVER.  However, there was an error in this
automatic definition, since it caused the TAP test driver to be
unconditionally invoked with perl.  This wasn't an issue in most
situations, since perl is smart enough to re-execute a given script
with the proper interpreter if it sees a she-bang line that doesn't
seem to point to perl itself.  Still, there is no reason to do
something blatantly wrong even if our tools correct the dumb mistake
for us.  So fix the TEST_LOG_DRIVER definition.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>