compat: substitute '@mkdir_p@' as '$(MKDIR_P)' unconditionally
We had started to do so only when the obsolete AM_PROG_MKDIR_P variable
was used, but that has recently turned out not to be not good enough,
because Gettext 0.18.2 (the latest one at the moment of writing) is using
the modern macro AC_PROG_MKDIR_P in its '*.m4' files, but at the same time
is still using the obsolete make variable '$(mkdir_p)' in its Makefile
fragments.  This has been causing failures in real-world packages
bootstrapped with latest Gettext (0.18.2) and with bleeding-edge automake.

See for example:

  <http://lists.gnu.org/archive/html/bug-tar/2013-04/msg00009.html>

And the worst things is that those failures have been basically silent
and unnoticeable up until the final "make install" --- that is, after
the affected package had been bootstrapped, configured and built.  Yikes!

This issue has been reported to the Gettext developers:

  <http://lists.gnu.org/archive/html/bug-gettext/2013-04/msg00044.html>

But even if they will quickly fix it, Gettext 0.18.2 will remain
affected.

The fix we can implement in Automake is small and unobtrusive enough,
definitely worth having in order to save the users and developers from
having to deal with this hard-to-spot issue.

* m4/init.m4 (AM_INIT_AUTOMAKE): AC_SUBST '@mkdir_p@' to '$(MKDIR_P)'.
* t/am-prog-mkdir-p.sh: Enhance.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>