refactor: fix few "inverted boolean" usages
In some subroutines, we used a return value of 0 to indicate success,
and a return status of 1 to indicate failure.  That was not very
consistent with the perl interpretation of 0 as a false value and 1 as
a true value.  So we now invert the meaning of the exit statuses.

* lib/Automake/Options.pm (_process_option_list): Here.
(process_global_option_list, process_option_list): And by reflex,
here as well.
* bin/automake.in (handle_options): And here.
(generate_makefile, scan_autoconf_traces): Adjust.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>