tests: remove remaining exec bits ('master' branch)
The executable bit gives the impression that the tests are directly
runnable, as with "./t/foo.sh", but it has been a while since that
was the case.  Today, tests are runnable only through "make check"
or "./runtest".

This change is for the 'master' branch (automake 1.99a), and is a
follow-up to commit 'v1.13.2-3-g74017b5', done on the 'micro' branch
(automake 1.13.2a) and to commit 'v1.13.2-221-g0985d67', done on the
'maint' branch (automake 1.13a).

* t/am-prog-mkdir-p.sh: Remove executable bit.
* t/txinfo-no-split.sh: Likewise.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>