texi: build version.texi and stamp-vti in srcdir
Do so even when the 'info-in-builddir' option is present, or when
the corresponding '*.info' files are listed in $(CLEANFILES) or in
$(DISTCLEANFILES).

This fixes failures in the following tests, when they are run with
$MAKE pointing to FreeBSD make:

    - txinfo-nodist-info.sh
    - txinfo23.sh
    - txinfo24.sh
    - txinfo28.sh
    - txinfo25.sh

BTW, notice that the test 'txinfo-builddir.sh' fails with FreeBSD make
as well, but that is due to a known FreeBSD make VPATH issue (the same
described in automake bug#7884).  But that is not a regression, since
the 'info-in-builddir' option will be new in Automake 1.14.  Moreover,
we already warn in the manual that the use of that option can indeed
cause problems with VPATH builds done by non-GNU make.

* bin/automake.in (handle_texinfo_helper): New local variable
'$soutdir'.  Use it instead of '$outdir' where appropriate (in
particular, in the transform used on file 'texi-vers.am'.
* t/txinfo-builddir.sh: Adjust to avoid spurious failures.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>