tests: use append mode to capture parallel make output
This will allow us to run the Automake testsuite forcing all the make
invocations in the test cases to run in parallel mode [1], but without
hitting the spurious failure reported in automake bug#11413.

See also:
<http://lists.gnu.org/archive/html/bug-make/2013-05/msg00135.html>

The make invocations in the test cases can be forced to run in parallel
mode by, e.g., invoking "make check" like this:

      make check AM_TESTSUITE_MAKE="make -j16"

The possible spurious failures hinted above are due to the fact that
the redirected output of parallel make can racily lose lines.  For
example, if GNU make (3.82) is run with -j10 on a Makefile like this:

    all = 0 1 2 3 4 5 6 7 8 9
    default: $(all)
    $(all):
            @sleep 0.$$(($RANDOM % 10)); echo $@

and has its standard output redirected in overwrite-mode to a regular
file, it looses a line of that output every 15 runs or so on a Fedora
17 system with 64 ppc64 cores.  Redirection in append mode does not
suffer of this problem, as explained here:
<http://lists.gnu.org/archive/html/automake-patches/2009-03/msg00073.html>

See also previous commits v1.12-63-g45c1fcd of 2012-05-05 (tests: use
append mode to capture parallel make output) and Release-1-10-280-g6426999
of 2009-03-10 (Use append mode to capture parallel test output).

* t/ax/am-test-lib.sh (run_make): Adjust and enhance.
* t/lisp8.sh: Take advantage of the enhancement, nd stop doing output
redirection for $MAKE by hand.
* t/tap-more.sh: Likewise.
* t/parallel-tests-concurrency.sh: Likewise.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>