announcement: can be generated from development snapshots as well
And not only from checkouts corresponding exactly from a beta
or stable release.  That was only getting in the way of proper
testing for the 'announcement' recipe.

* maint.mk (determine_release_type): If the make variable
DEVEL_SNAPSHOT is set, do not error out if the current version
denotes a development snapshot (e.g., "1.13.2c" or "1.99a").
(announcement): Relax, by also accepting to run from development
snapshots, not only stable or beta releases.  Do so by defining
the target-specific variable DEVEL_SNAPSHOT to "yes".

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>